requirements.txt
flake8
isort
mypy
black
python-dotenv

aws-cdk.core==1.104.0
aws-cdk.aws-cognito==1.104.0



